ROE-2309 Separated update and register resume
