+++ src/controllers/shared/common.resume.submission.controller.ts
@@ -1,14 +1,10 @@
-import { NextFunction, Request, Response } from "express";
 import { v4 as uuidv4 } from "uuid";
 import { Session } from "@companieshouse/node-session-handler";
 
 import * as config from "../../config";
-
 import { ApplicationData } from "../../model";
-import { createAndLogErrorRequest, logger } from "../../utils/logger";
 import { setExtraData } from "../../utils/application.data";
 import { isActiveFeature } from "../../utils/feature.flag";
-import { getOverseasEntity } from "../../service/overseas.entities.service";
 
 import { HasSoldLandKey, ID, IsSecureRegisterKey, OverseasEntityKey, Transactionkey } from "../../model/data.types.model";
 import { WhoIsRegisteringKey, WhoIsRegisteringType } from "../../model/who.is.making.filing.model";
@@ -20,48 +16,7 @@ import { BeneficialOwnerOther, BeneficialOwnerOtherKey } from "../../model/benef
 import { ManagingOfficerCorporate, ManagingOfficerCorporateKey } from "../../model/managing.officer.corporate.model";
 import { ManagingOfficerIndividual, ManagingOfficerKey } from "../../model/managing.officer.model";
 
-import { startPaymentsSession } from "../../service/payment.service";
-import { getTransaction } from "../../service/transaction.service";
 import { mapTrustApiReturnModelToWebModel } from "../../utils/trusts";
-
-export const getResumePage = async (req: Request, res: Response, next: NextFunction, resumePage: string) => {
-  try {
-    logger.debugRequest(req, `GET a saved OE submission`);
-
-    const { transactionId, overseaEntityId } = req.params;
-    const infoMsg = `Transaction ID: ${transactionId}, OverseasEntity ID: ${overseaEntityId}`;
-
-    logger.infoRequest(req, `Resuming OE - ${infoMsg}`);
-    const appData: ApplicationData = await getOverseasEntity(req, transactionId, overseaEntityId);
-
-    if (!Object.keys(appData || {}).length) {
-      throw createAndLogErrorRequest(req, `Error on resuming OE - ${infoMsg}`);
-    }
-
-    const session = req.session as Session;
-    setWebApplicationData(session, appData, transactionId, overseaEntityId);
-
-    const transactionResource = await getTransaction(req, transactionId);
-
-    if (transactionResource.status === config.CLOSED_PENDING_PAYMENT) {
-      const headersPaymentUrl = {
-        headers: {
-          [config.PAYMENT_REQUIRED_HEADER]: config.PAYMENTS_API_URL + config.PAYMENTS
-        }
-      };
-      const redirectPath = await startPaymentsSession(req, session, transactionId, overseaEntityId, headersPaymentUrl);
-
-      logger.infoRequest(req, `Payments Session created on Resume link with, Trans_ID: ${transactionId}, OE_ID: ${overseaEntityId}. Redirect to: ${redirectPath}`);
-
-      return res.redirect(redirectPath);
-    }
-
-    return res.redirect(resumePage);
-  } catch (error) {
-    logger.errorRequest(req, error);
-    next(error);
-  }
-};
 
 /**
  * Set default values needed for the web journey that are not part of OE API data model
@@ -73,7 +28,7 @@ export const getResumePage = async (req: Request, res: Response, next: NextFunct
  * @param transactionId
  * @param overseaEntityId
  */
-const setWebApplicationData = (session: Session, appData: ApplicationData, transactionId: string, overseaEntityId: string) => {
+export const setWebApplicationData = (session: Session, appData: ApplicationData, transactionId: string, overseaEntityId: string) => {
 
   appData[BeneficialOwnerIndividualKey] = (appData[BeneficialOwnerIndividualKey] as BeneficialOwnerIndividual[])
     .map( boi => { return { ...boi, [ID]: uuidv4() }; } );
